fix: activity card height && refactored: moved activity card for a separate file
