+++ src/components/Dashboard/Activities/ActivityColumn.js
@@ -2,13 +2,11 @@ import dayjs from "dayjs";
 import timezone from "dayjs/plugin/timezone";
 import utc from "dayjs/plugin/utc";
 import {
-  ActivityCard,
-  ActivityInfoContainer,
   ColumnActivities,
   ContainerColumnActivities,
   TitleColumnActivities,
 } from "./Styles";
-import SubscribeField from "./SubscribeField";
+import ActivityCard from "./ActivityCard";
 
 dayjs.extend(utc);
 dayjs.extend(timezone);
@@ -21,10 +19,6 @@ export default function ActivityColumn({
   title,
   location,
 }) {
-  const getActivityCardHeight = (startDate, endDate) => {
-    return dayjs(endDate).diff(startDate, "hours", true) * 80;
-  };
-
   return (
     <ContainerColumnActivities>
       <TitleColumnActivities>{title}</TitleColumnActivities>
@@ -34,18 +28,11 @@ export default function ActivityColumn({
           .map(({ id, name, startDate, endDate, availableVacancies }) => (
             <ActivityCard
               key={id}
-              $height={() => getActivityCardHeight(startDate, endDate)}
-            >
-              <ActivityInfoContainer>
-                <p>{name}</p>
-                <span>
-                  {dayjs(startDate).format("HH:mm") +
-                    " - " +
-                    dayjs(endDate).format("HH:mm")}
-                </span>
-              </ActivityInfoContainer>
-              <SubscribeField availableVacancies={availableVacancies} />
-            </ActivityCard>
+              name={name}
+              startDate={startDate}
+              endDate={endDate}
+              availableVacancies={availableVacancies}
+            />
           ))}
       </ColumnActivities>
     </ContainerColumnActivities>
